Move all tipping functionality to its own module
